2023-11-27, sse 1차 완성, 공부중

+++ server-sent-event/src/main/java/com/example/serversentevent/dto/RequestUpdateStageOpenDto.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class RequestUpdateStageOpenDto {
-    private Long num;
+    private Integer num;
     private String str;
 }
