+++ server-sent-event/src/main/java/com/example/serversentevent/repository/SseRepository.java
@@ -3,6 +3,6 @@ package com.example.serversentevent.repository;
 import java.util.Optional;
 
 public interface SseRepository {
-    Optional<String> find(Long num);
-    void update(Long num, String str);
+    Optional<String> find(Integer num);
+    void update(Integer num, String str);
 }
