+++ server-sent-event/src/main/java/com/example/serversentevent/service/SseService.java
@@ -3,6 +3,6 @@ package com.example.serversentevent.service;
 import com.example.serversentevent.dto.RequestUpdateStageOpenDto;
 
 public interface SseService {
-    String findStageOpen(Long num);
+    String findStageOpen(Integer num);
     void updateStageOpen(RequestUpdateStageOpenDto dto);
 }
